+++ lib/blog.ex
@@ -1,4 +1,6 @@
 defmodule Blog do
   @moduledoc "For testing only"
-  use Postex, prefix: "https://postex.com/posts/"
+  use Postex,
+    prefix: "https://postex.com/posts/",
+    per_page: 1
 end